+++ app/Plugin/Schedules.php
@@ -37,10 +37,11 @@ class Schedules {
 	 * Return the instance of this Singleton object.
 	 */
 	public static function get_instance(): Schedules {
-		if ( ! static::$instance instanceof static ) {
-			static::$instance = new static();
+		if ( is_null( self::$instance ) ) {
+			self::$instance = new self();
 		}
-		return static::$instance;
+
+		return self::$instance;
 	}
 
 	/**
@@ -76,7 +77,7 @@ class Schedules {
 	/**
 	 * Get our own active events from WP-list.
 	 *
-	 * @return array
+	 * @return array<string,array<string,mixed>>
 	 */
 	private function get_events(): array {
 		// get our own events from events list in WordPress.
@@ -88,7 +89,7 @@ class Schedules {
 		 *
 		 * @since 2.0.0 Available since 2.0.0.
 		 *
-		 * @param array $our_events List of our own events in WP-cron.
+		 * @param array<string,array<string,mixed>> $our_events List of our own events in WP-cron.
 		 */
 		return apply_filters( 'product_swatches_light_schedule_our_events', $our_events );
 	}
@@ -100,9 +101,9 @@ class Schedules {
 	 *
 	 * Does only run in wp-admin, not frontend.
 	 *
-	 * @param array $our_events List of our own events.
+	 * @param array<string,array<string,mixed>> $our_events List of our own events.
 	 *
-	 * @return array
+	 * @return array<string,array<string,mixed>>
 	 */
 	public function check_events( array $our_events ): array {
 		// bail if check should be disabled.
@@ -178,7 +179,7 @@ class Schedules {
 	/**
 	 * Return list of all schedule-object-names.
 	 *
-	 * @return array
+	 * @return array<int,string>
 	 */
 	public function get_schedule_object_names(): array {
 		// list of schedules: free version supports only one import-schedule.
@@ -193,7 +194,7 @@ class Schedules {
 		 *
 		 * @since 2.0.0 Available since 2.0.0.
 		 *
-		 * @param array $list_of_schedules List of additional schedules.
+		 * @param array<int,string> $list_of_schedules List of additional schedules.
 		 */
 		return apply_filters( 'product_swatches_light_schedules', $list_of_schedules );
 	}
@@ -218,7 +219,7 @@ class Schedules {
 	/**
 	 * Get our own events from WP-cron-event-list.
 	 *
-	 * @return array
+	 * @return array<string,array<string,mixed>>
 	 */
 	private function get_wp_events(): array {
 		$our_events = array();
@@ -251,10 +252,10 @@ class Schedules {
 		}
 
 		// get our object.
-		$schedule_obj = $this->get_schedule_object_by_name( $event->hook );
+		$schedule_obj = $this->get_schedule_object_by_name( $event->hook ); // @phpstan-ignore property.notFound
 
 		// bail if this is not an event of our plugin.
-		if ( ! $schedule_obj ) {
+		if ( ! $schedule_obj instanceof Schedules_Base ) {
 			return $event;
 		}