+++ app/Swatches/AttributeType.php
@@ -7,6 +7,11 @@
 
 namespace ProductSwatchesLight\Swatches;
 
+// prevent direct access.
+defined( 'ABSPATH' ) || exit;
+
+use WP_Term;
+
 /**
  * Interface for each attribute-type this plugin supports.
  */
@@ -14,13 +19,13 @@ interface AttributeType {
 	/**
 	 * Output on list page.
 	 *
-	 * @param array  $item_list The list.
-	 * @param array  $images The images.
-	 * @param array  $images_sets The image sets.
-	 * @param array  $values The values.
-	 * @param array  $on_sales The sales markers.
-	 * @param string $product_link The product URL.
-	 * @param string $product_title The product title.
+	 * @param array<int,WP_Term>   $item_list The list.
+	 * @param array<string,string> $images The images.
+	 * @param array<string,string> $images_sets The image sets.
+	 * @param array<string,mixed>  $values The values.
+	 * @param array<string,mixed>  $on_sales The sales markers.
+	 * @param string               $product_link The product URL.
+	 * @param string               $product_title The product title.
 	 * @return string
 	 */
 	public static function get_list( array $item_list, array $images, array $images_sets, array $values, array $on_sales, string $product_link, string $product_title ): string;
@@ -28,8 +33,8 @@ interface AttributeType {
 	/**
 	 * Output on taxonomy table in backend.
 	 *
-	 * @param int   $term_id The term id.
-	 * @param array $fields The fields.
+	 * @param int                 $term_id The term id.
+	 * @param array<string,mixed> $fields The fields.
 	 *
 	 * @return string
 	 */
@@ -38,9 +43,9 @@ interface AttributeType {
 	/**
 	 * Return values of a field with this attribute-type.
 	 *
-	 * @param int          $term_id The term id.
-	 * @param string|array $term The term settings.
-	 * @return array
+	 * @param int                        $term_id The term id.
+	 * @param string|array<string,mixed> $term The term settings.
+	 * @return array<string,mixed>
 	 */
 	public static function get_values( int $term_id, string|array $term ): array;
 }